+++ CHANGELOG.md
@@ -794,3 +794,4 @@
 - [2026-05-02 15:55] perf(theme): apply RepaintBoundary on animated theme widgets
 - [2026-05-02 16:22] feat(storage): add bi-directional event bus handler for storage
 - [2026-05-02 16:26] perf(state): apply RepaintBoundary on animated state widgets
+- [2026-05-02 17:10] feat(ui): implement responsive ui view component